fix: move test fixtures from integration to workflows conftest
The integration conftest was unreachable from the workflow tests
since pytest conftest fixtures are scoped to their directory and
below. Moved the library registration and state cleanup fixtures
to the workflows conftest where they are actually needed.
